+++ gluecodium/src/test/java/com/here/gluecodium/validator/LimeOverloadsValidatorTest.kt
@@ -20,16 +20,16 @@
 package com.here.gluecodium.validator
 
 import com.here.gluecodium.model.lime.LimeBasicTypeRef
-import com.here.gluecodium.model.lime.LimeClass
+import com.here.gluecodium.model.lime.LimeContainer
 import com.here.gluecodium.model.lime.LimeElement
 import com.here.gluecodium.model.lime.LimeFunction
 import com.here.gluecodium.model.lime.LimeParameter
 import com.here.gluecodium.model.lime.LimePath
 import com.here.gluecodium.model.lime.LimePath.Companion.EMPTY_PATH
 import com.here.gluecodium.model.lime.LimeSignatureResolver
-import com.here.gluecodium.model.lime.LimeStruct
 import io.mockk.mockk
 import org.junit.Assert.assertEquals
+import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
@@ -38,30 +38,32 @@ import org.junit.runners.Parameterized
 class LimeOverloadsValidatorTest(
     private val limeMethod1: LimeFunction,
     private val limeMethod2: LimeFunction,
-    private val expectedResult: Boolean
+    private val expectedResultNoConstructors: Boolean,
+    private val expectedResultWithConstructors: Boolean
 ) {
     private val allElements = mutableMapOf<String, LimeElement>()
 
-    private val validator = LimeOverloadsValidator(LimeSignatureResolver(allElements), mockk(relaxed = true))
+    private val validatorNoConstructors =
+        LimeOverloadsValidator(LimeSignatureResolver(allElements), mockk(relaxed = true), validateConstructors = false)
+    private val validatorWithConstructors =
+        LimeOverloadsValidator(LimeSignatureResolver(allElements), mockk(relaxed = true), validateConstructors = true)
 
-    @Test
-    fun validateInContainer() {
+    @Before
+    fun setUp() {
         allElements[limeMethod1.path.toString()] = limeMethod1
         allElements[limeMethod2.path.toString()] = limeMethod2
         allElements[fooPath.toString()] =
-            LimeClass(fooPath, functions = listOf(limeMethod1, limeMethod2))
-
-        assertEquals(expectedResult, validator.validate(allElements.values))
+            object : LimeContainer(fooPath, functions = listOf(limeMethod1, limeMethod2)) {}
     }
 
     @Test
-    fun validateInStruct() {
-        allElements[limeMethod1.path.toString()] = limeMethod1
-        allElements[limeMethod2.path.toString()] = limeMethod2
-        allElements[fooPath.toString()] =
-            LimeStruct(fooPath, functions = listOf(limeMethod1, limeMethod2))
+    fun validateNoConstructors() {
+        assertEquals(expectedResultNoConstructors, validatorNoConstructors.validate(allElements.values))
+    }
 
-        assertEquals(expectedResult, validator.validate(allElements.values))
+    @Test
+    fun validateWithConstructors() {
+        assertEquals(expectedResultWithConstructors, validatorWithConstructors.validate(allElements.values))
     }
 
     companion object {
@@ -75,11 +77,43 @@ class LimeOverloadsValidatorTest(
             arrayOf(
                 LimeFunction(barPath.withSuffix("0")),
                 LimeFunction(barPath.withSuffix("1"), parameters = listOf(limeParameter)),
+                true,
                 true
             ),
             arrayOf(
                 LimeFunction(barPath.withSuffix("0"), parameters = listOf(limeParameter)),
                 LimeFunction(barPath.withSuffix("1"), parameters = listOf(limeParameter)),
+                false,
+                false
+            ),
+            arrayOf(
+                LimeFunction(barPath.withSuffix("0"), isConstructor = true),
+                LimeFunction(barPath.withSuffix("1"), isConstructor = true, parameters = listOf(limeParameter)),
+                true,
+                true
+            ),
+            arrayOf(
+                LimeFunction(barPath.withSuffix("0"), isConstructor = true, parameters = listOf(limeParameter)),
+                LimeFunction(barPath.withSuffix("1"), isConstructor = true, parameters = listOf(limeParameter)),
+                false,
+                false
+            ),
+            arrayOf(
+                LimeFunction(barPath.withSuffix("0")),
+                LimeFunction(barPath.withSuffix("1"), isConstructor = true, parameters = listOf(limeParameter)),
+                true,
+                true
+            ),
+            arrayOf(
+                LimeFunction(barPath.withSuffix("0"), parameters = listOf(limeParameter)),
+                LimeFunction(barPath.withSuffix("1"), isConstructor = true, parameters = listOf(limeParameter)),
+                false,
+                false
+            ),
+            arrayOf(
+                LimeFunction(barPath, isConstructor = true, parameters = listOf(limeParameter)),
+                LimeFunction(fooPath.child("wee"), isConstructor = true, parameters = listOf(limeParameter)),
+                true,
                 false
             )
         )